requirements.txt
python-telegram-bot==21.6
Flask==3.0.3
waitress==3.0.0
yt-dlp==2025.01.15
requests==2.32.3
segno==1.6.1
pypng==0.20220715.0
